requirements.txt
py-trello==0.19.0
python-dotenv==1.0.0
flask==3.0.0
